decoupled hygiene and ScopeAnalyzer
